refactor: 프로모션 관련 타이머 및 핸들러를 별도 서비스 파일로 분리
